merge to agustin

+++ RevStay/src/main/java/com/Rev/RevStay/services/RoomService.java
@@ -1,16 +1,18 @@
 package com.Rev.RevStay.services;
 
 import com.Rev.RevStay.exceptions.GenericException;
-import com.Rev.RevStay.models.Hotel;
 import com.Rev.RevStay.models.Room;
 
+import com.Rev.RevStay.models.User;
 import com.Rev.RevStay.repos.HotelDAO;
 import com.Rev.RevStay.repos.RoomDAO;
+import com.Rev.RevStay.repos.UserDAO;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.math.BigDecimal;
+import java.nio.file.AccessDeniedException;
 import java.util.List;
 import java.util.Optional;
 
@@ -19,14 +21,17 @@ import java.util.Optional;
 public class RoomService {
     private final RoomDAO roomDAO;
     private final HotelDAO hotelDAO;
+    private final UserDAO userDAO;
+
 
     @Autowired
-    public RoomService(RoomDAO roomDAO, HotelDAO hotelDAO) { this.roomDAO = roomDAO;
+    public RoomService(RoomDAO roomDAO, HotelDAO hotelDAO, UserDAO userDAO) { this.roomDAO = roomDAO;
         this.hotelDAO = hotelDAO;
+        this.userDAO = userDAO;
     }
 
     //TODO Register New ROOM
-    public Optional<Room> register(Room roomToBeCreate){
+    public Optional<Room> register(Room roomToBeCreate, int ownerId){
 
         if (roomToBeCreate.getHotel() == null || roomToBeCreate.getHotel().getHotelId() == 0) {
             throw new GenericException("Hotel information is required");
@@ -42,48 +47,62 @@ public class RoomService {
             throw new GenericException("The number of guests must be at least one");
         }
 
-        return Optional.of(roomDAO.save(roomToBeCreate));
+        Optional<User> owner = userDAO.findById(ownerId);
+        if (owner.isPresent()){
+            if (roomToBeCreate.getHotel().getOwner() == owner.get()) {
+                return Optional.of(roomDAO.save(roomToBeCreate));
+            } else {
+                throw new GenericException("You are not authorized to register rooms in this Hotel.");
+            }
+        }
+        return Optional.empty();
     }
 
     //TODO Delete ROOM
-    public void deleteRoom(int roomId){
+    public void deleteRoom(int roomId, int ownerId){
         Room roomToBeDelete = roomDAO.findById(roomId)
                 .orElseThrow(() -> new GenericException("Room not found"));
-
-        roomDAO.delete(roomToBeDelete);
+        Optional<User> owner = userDAO.findById(ownerId);
+        if (owner.isPresent()){
+            if (roomToBeDelete.getHotel().getOwner() == owner.get()) {
+                roomDAO.delete(roomToBeDelete);
+            } else {
+                throw new GenericException("You are not authorized to delete this room.");
+            }
+        }
     }
 
     // TODO Update ROOM
-    public Optional<Room> updateRoom(int roomId, Room updatedRoom) {
+    public Optional<Room> updateRoom(int roomId, Room updatedRoom, int ownerId) {
         Room existingRoom = roomDAO.findById(roomId)
                 .orElseThrow(() -> new GenericException("Room not found"));
+        Optional<User> owner = userDAO.findById(ownerId);
+        if (owner.isPresent()){
+            if (existingRoom.getHotel().getOwner() == owner.get()) {
 
-        // The hotel is not the same ??
-        if (updatedRoom.getHotel() != null) {
-            int hotelId = updatedRoom.getHotel().getHotelId();
-            if (!hotelDAO.existsById(hotelId)) {
-                throw new GenericException("The hotel doesn't exist");
+                if (updatedRoom.getPrice() != null && updatedRoom.getPrice().compareTo(BigDecimal.ZERO) > 0) {
+                    existingRoom.setPrice(updatedRoom.getPrice());
+                }
+
+                if (updatedRoom.getMaxGuests() > 0) {
+                    existingRoom.setMaxGuests(updatedRoom.getMaxGuests());
+                }
+
+                if (updatedRoom.getDescription() != null) {
+                    existingRoom.setDescription(updatedRoom.getDescription());
+                }
+
+                if (updatedRoom.getRoomType() != null) {
+                    existingRoom.setRoomType(updatedRoom.getRoomType());
+                }
+
+                return Optional.of(roomDAO.save(existingRoom));
+
+            } else {
+                throw new GenericException("You are not authorized to modify this room.");
             }
-            existingRoom.setHotel(updatedRoom.getHotel());
         }
-
-        if (updatedRoom.getPrice() != null && updatedRoom.getPrice().compareTo(BigDecimal.ZERO) > 0) {
-            existingRoom.setPrice(updatedRoom.getPrice());
-        }
-
-        if (updatedRoom.getMaxGuests() > 0) {
-            existingRoom.setMaxGuests(updatedRoom.getMaxGuests());
-        }
-
-        if (updatedRoom.getDescription() != null) {
-            existingRoom.setDescription(updatedRoom.getDescription());
-        }
-
-        if (updatedRoom.getRoomType() != null) {
-            existingRoom.setRoomType(updatedRoom.getRoomType());
-        }
-
-        return Optional.of(roomDAO.save(existingRoom));
+        return Optional.empty();
     }
 
     //TODO Get Rooms By Hotel Id
